backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
scikit-learn==1.5.0
tensorflow==2.16.1
numpy==1.26.4
pandas==2.2.2
pillow==10.3.0
torch
torchvision
pyyaml
python-multipart
aiofiles
